requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-countries
async-timeout==4.0.2
    # via redis
beautifulsoup4==4.12.2
    # via -r requirements.in
cachetools==4.2.4
    # via google-auth
certifi==2020.11.8
    # via requests
charset-normalizer==2.0.3
    # via requests
cssmin==0.2.0
    # via -r requirements.in
cssselect2==0.4.1
    # via svglib
django==4.2.3
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-csp
    #   django-filer
    #   django-formtools
    #   django-haystack
    #   django-pagedown
    #   django-polymorphic
    #   django-ranged-response
    #   django-redis
    #   django-reversion
    #   django-sekizai
    #   django-simple-captcha
    #   django-storages
    #   django-treebeard
    #   easy-thumbnails
    #   jsonfield
django-appconf==1.0.4
    # via django-compressor
django-classy-tags==3.0.1
    # via
    #   django-cms
    #   django-sekizai
django-cms==3.11.3
    # via
    #   -r requirements.in
    #   djangocms-attributes-field
    #   djangocms-file
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-text-ckeditor
django-compressor==4.4
    # via -r requirements.in
django-countries==7.5.1
    # via -r requirements.in
django-csp==3.7
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-filer==2.2.6
    # via
    #   -r requirements.in
    #   djangocms-file
    #   djangocms-link
    #   djangocms-picture
django-formtools==2.4.1
    # via django-cms
django-haystack==3.2.1
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-markdown-deux==1.0.6
    # via -r requirements.in
django-mptt==0.14.0
    # via
    #   -r requirements.in
    #   django-filer
django-pagedown==2.2.1
    # via -r requirements.in
django-polymorphic==3.0.0
    # via django-filer
django-ranged-response==0.2.0
    # via django-simple-captcha
django-redis==5.3.0
    # via -r requirements.in
django-reversion==5.0.4
    # via -r requirements.in
django-sekizai==4.1.0
    # via
    #   -r requirements.in
    #   django-cms
django-simple-captcha==0.5.18
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
django-treebeard==4.7
    # via
    #   -r requirements.in
    #   django-cms
djangocms-admin-style==3.2.4
    # via
    #   -r requirements.in
    #   django-cms
djangocms-attributes-field==3.0.0
    # via
    #   -r requirements.in
    #   djangocms-file
    #   djangocms-link
    #   djangocms-picture
djangocms-file==3.0.1
    # via -r requirements.in
djangocms-link==3.1.1
    # via -r requirements.in
djangocms-picture==4.0.0
    # via -r requirements.in
djangocms-text-ckeditor==5.1.3
    # via -r requirements.in
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-picture
feedparser==6.0.10
    # via -r requirements.in
google-api-core==2.8.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.21.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.0
    # via google-cloud-storage
google-cloud-storage==2.10.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via google-cloud-storage
googleapis-common-protos==1.56.2
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
html5lib==1.1
    # via djangocms-text-ckeditor
idna==2.10
    # via requests
iso3166==2.1.1
    # via -r requirements.in
jsonfield==3.1.0
    # via -r requirements.in
libsass==0.22.0
    # via -r requirements.in
lxml==4.6.5
    # via svglib
markdown2==2.4.0
    # via django-markdown-deux
micawber==0.5.5
    # via -r requirements.in
packaging==21.3
    # via
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.0.0
    # via
    #   -r requirements.in
    #   django-pagedown
    #   django-simple-captcha
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
protobuf==3.19.1
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via packaging
rcssmin==1.1.1
    # via django-compressor
redis==4.5.3
    # via django-redis
reportlab==3.6.5
    # via
    #   easy-thumbnails
    #   svglib
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rjsmin==1.2.1
    # via django-compressor
rsa==4.7.2
    # via google-auth
sgmllib3k==1.0.0
    # via feedparser
six==1.15.0
    # via
    #   google-auth
    #   html5lib
soupsieve==2.0.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
svglib==1.1.0
    # via easy-thumbnails
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.2.0
    # via
    #   asgiref
    #   django-countries
unidecode==1.1.1
    # via django-filer
urllib3==1.26.5
    # via
    #   google-auth
    #   requests
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whoosh==2.7.4
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
